Seeders use default profile picture

+++ app/Http/Controllers/Api/InstructorController.php
@@ -137,7 +137,10 @@ class InstructorController extends Controller
                 $request->file('profile_picture'),
                 $account->username . '-' . time() . '.' . $request->profile_picture->extension()
             );
-            Storage::delete($account->profile_picture);
+            if ($account->profile_picture != "user_default.svg")
+            {
+                Storage::delete($account->profile_picture);
+            }
         }
 
         /** Hashify the password */
@@ -178,7 +181,10 @@ class InstructorController extends Controller
         $account = Instructor::find($data['id']);
 
         /** Delete the profile picture file. */
-        Storage::delete($account->profile_picture);
+        if ($account->profile_picture != "user_default.svg")
+        {
+            Storage::delete($account->profile_picture);
+        }
 
         /** Delete the account from database. */
         $account->delete();
